[VSPC-276] Biblioblock uses APACitationFormatter to show crroect formats
